Create robot_for_model_v0.2.py
When we do Device model QA, we need to split the .mdl and .lib file of our total model into several different couple of .lib file and .mdl file(for different colleagues to do model fitting for different MOSFET like PB, NMOS, PMOS.....).   This v0.2 version is used for this purpose, split the BIG total model.